add unit test.

+++ src/SingleReadProcessorWithUmi.cpp
@@ -45,7 +45,7 @@ namespace barcodeSpace {
         Sequence previous = read.subRead(subReadsBoundaries.front().first, subReadsBoundaries.front().second);
         // If the umi region is out of read index.
         // exit the program
-        if (read.length() < subReadsBoundaries.back().first) {
+        if (static_cast<int>(read.length()) < subReadsBoundaries.back().first) {
             std::cerr << "The umi region [" << subReadsBoundaries.front().second << ',' << subReadsBoundaries.back().first << ']' << " exceeds the read length " << read.length() << std::endl;
             throw new runtime_error("Incompatible umi region with read length!");
         }
